feat: Add Docker to the application
During development, it is necessary to run the application locally and create a database to connect it to the API. To facilitate this process, Docker was implemented in the application.

This commit adds Docker to the application, which allows the developer to run a container that runs the database and API through the npm run docker commands.